discord_bot/pyproject.toml
[tool.poetry]
authors = []
description = ""
name = "hogehoge"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {extras = ["voice"], version = "^1.7.3"}
opencv-python = "^4.5.5"
openai = "^0.18.1"
google-cloud-texttospeech = "^2.11.0"
wolframalpha = "^5.0.0"
Janome = "^0.4.2"
sympy = "^1.10.1"
mojimoji = "^0.0.12"
sentencepiece = "^0.1.96"

[tool.poetry.dev-dependencies]
pysen = {extras = ["lint"], version = "^0.10.1"}
pytest = "^6.2.4"
click = "8.0.2"

[tool.poetry.scripts]
server = 'hogehoge.main:server'

[tool.black] # automatically generated by pysen
line-length = 100
target-version = ["py37"]

[tool.isort] # automatically generated by pysen
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 100
multi_line_output = 3
use_parentheses = true
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pysen]
version = "0.10"

[tool.pysen-cli]
settings_dir = ""

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
line_length = 100
mypy_preset = "strict"
py_version = "py37"

[[tool.pysen.lint.mypy_targets]]
paths = [
  "./src",
]

[tool.pysen.lint.source]
includes = [
  ".",
]
# 無視するフォルダー
excludes = [
  # ex. "./src/xxx"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
